fix: move cmds to src and remove commit-lint
