+++ src/cmds/z-lint-package-json.js
@@ -2,8 +2,10 @@
 
 const execa = require('execa')
 const path = require('path')
-const { fromAegir, fromRoot } = require('../src/utils')
-
+const { fromAegir, fromRoot } = require('../utils')
+/**
+ * @typedef {import("yargs").Argv} Argv
+ */
 const EPILOG = `
 Supports options forwarding with '--' for more info check https://github.com/tclindner/npm-package-json-lint#cli-commands-and-configuration
 `
@@ -12,10 +14,16 @@ module.exports = {
   command: 'lint-package-json',
   desc: 'Lint package.json with aegir defaults.',
   aliases: ['lint-package', 'lpj'],
+  /**
+   * @param {Argv} yargs
+   */
   builder: (yargs) => {
     yargs
       .epilog(EPILOG)
   },
+  /**
+   * @param {{ [x: string]: any; _: string | any[]; }} argv
+   */
   handler (argv) {
     const input = argv._.slice(1)
     const forwardOptions = argv['--'] ? argv['--'] : []
